Replace FloatBuffer method in AudioSource/Listener classes; 3D audio test;

+++ src/com/base/engine/rendering/Camera.java
@@ -6,8 +6,7 @@ import com.base.engine.math.Matrix4f;
 import com.base.engine.math.Vector2f;
 import com.base.engine.math.Vector3f;
 
-public class Camera
-{
+public class Camera {
 
 	public static final Vector3f yAxis = new Vector3f(0, 1, 0);
 
@@ -16,23 +15,20 @@ public class Camera
 	private Vector3f up;
 	private Matrix4f projection;
 
-	public Camera(float fov, float aspectRatio, float zNear, float zFar)
-	{
+	public Camera(float fov, float aspectRatio, float zNear, float zFar) {
 		this.pos = new Vector3f(0, 0, 0);
 		this.forward = new Vector3f(0, 0, 1).normalized();
 		this.up = new Vector3f(0, 1, 0).normalized();
 		this.projection = new Matrix4f().initPerspective(fov, aspectRatio, zNear, zFar);
 	}
 
-	public Camera(Vector3f pos, Vector3f forward, Vector3f up)
-	{
+	public Camera(Vector3f pos, Vector3f forward, Vector3f up) {
 		this.pos = pos;
 		this.forward = forward.normalized();
 		this.up = up.normalized();
 	}
 
-	public Matrix4f getViewProjection()
-	{
+	public Matrix4f getViewProjection() {
 		Matrix4f cameraRotation = new Matrix4f().initRotation(forward, up);
 		Matrix4f cameraTranslation = new Matrix4f().initTranslation(-pos.getX(), -pos.getY(), -pos.getZ());
 		return projection.mul(cameraRotation.mul(cameraTranslation));
@@ -41,28 +37,26 @@ public class Camera
 	boolean mouseLocked = false;
 	Vector2f centerPosition = new Vector2f(Window.getWidth() / 2, Window.getHeight() / 2);
 
-	public void input(float delta)
-	{
+	public void input(float delta) {
 		float sensitivity = 0.1f;
-		float movAmt = (float) (10 * delta);
+		float movAmt = 10 * delta;
 		float rotAmt = (float) (100 * Time.getDelta());
 
-		if(Input.getKey(Input.KEY_ESCAPE))
-		{
+		if (Input.getKey(Input.KEY_ESCAPE)) {
 			Input.SetCursor(true);
 			mouseLocked = false;
 		}
 
-		if(Input.getKey(Input.KEY_W))
+		if (Input.getKey(Input.KEY_W))
 			move(getForward(), movAmt);
-		if(Input.getKey(Input.KEY_S))
+		if (Input.getKey(Input.KEY_S))
 			move(getForward(), -movAmt);
-		if(Input.getKey(Input.KEY_A))
+		if (Input.getKey(Input.KEY_A))
 			move(getLeft(), movAmt);
-		if(Input.getKey(Input.KEY_D))
+		if (Input.getKey(Input.KEY_D))
 			move(getRight(), movAmt);
 
-		if(Input.GetMouseDown(0))
+		if (Input.GetMouseDown(0))
 
 		{
 			Input.SetMousePosition(centerPosition);
@@ -70,20 +64,19 @@ public class Camera
 			mouseLocked = true;
 		}
 
-		if(mouseLocked)
-		{
+		if (mouseLocked) {
 
 			Vector2f deltaPos = Input.GetMousePosition().sub(centerPosition);
 
 			boolean rotY = deltaPos.getX() != 0;
 			boolean rotX = deltaPos.getY() != 0;
 
-			if(rotY)
+			if (rotY)
 				rotateY(deltaPos.getX() * sensitivity);
-			if(rotX)
+			if (rotX)
 				rotateX(deltaPos.getY() * sensitivity);
 
-			if(rotY || rotX)
+			if (rotY || rotX)
 				Input.SetMousePosition(new Vector2f(Window.getWidth() / 2, Window.getHeight() / 2));
 		}
 
@@ -96,13 +89,11 @@ public class Camera
 
 	}
 
-	public void move(Vector3f dir, float amt)
-	{
+	public void move(Vector3f dir, float amt) {
 		pos = pos.add(dir.mul(amt));
 	}
 
-	public void rotateY(float angle)
-	{
+	public void rotateY(float angle) {
 		Vector3f Haxis = yAxis.cross(forward).normalized();
 
 		forward.rotate(angle, yAxis).normalized();
@@ -111,8 +102,7 @@ public class Camera
 
 	}
 
-	public void rotateX(float angle)
-	{
+	public void rotateX(float angle) {
 		Vector3f Haxis = yAxis.cross(forward).normalized();
 
 		forward.rotate(angle, Haxis).normalized();
@@ -121,45 +111,37 @@ public class Camera
 
 	}
 
-	public Vector3f getLeft()
-	{
+	public Vector3f getLeft() {
 		Vector3f left = forward.cross(up).normalized();
 		return left;
 	}
 
-	public Vector3f getRight()
-	{
+	public Vector3f getRight() {
 		Vector3f right = up.cross(forward).normalized();
 		return right;
 	}
 
-	public Vector3f getPos()
-	{
+	public Vector3f getPos() {
 		return pos;
 	}
 
-	public void setPos(Vector3f pos)
-	{
+	public void setPos(Vector3f pos) {
 		this.pos = pos;
 	}
 
-	public Vector3f getForward()
-	{
+	public Vector3f getForward() {
 		return forward;
 	}
 
-	public void setForward(Vector3f forward)
-	{
+	public void setForward(Vector3f forward) {
 		this.forward = forward;
 	}
 
-	public Vector3f getUp()
-	{
+	public Vector3f getUp() {
 		return up;
 	}
 
-	public void setUp(Vector3f up)
-	{
+	public void setUp(Vector3f up) {
 		this.up = up;
 	}
